requirements.txt
flask
openai-whisper
audio_extract